rpk: Add 'rpk redpanda config', deprecate 'rpk config'
